+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> MySQL components for Python Changelog
 
+## <a name="3.2.1"></a> 3.2.1 (2021-08-25)
+
+### Features
+* Added _request method for PostgresPersistence
 
 ## <a name="3.2.0"></a> 3.2.0 (2021-08-09)
 
 Added support for database schemas
-
+t push
 ### Features
 * Added schemas to MySqlPersistence, IdentifiableMySqlPersistence, IdentifiableJsonMySqlPersistence
 * Added _auto_generate_id flag to IdentifiableMySqlPersistence